Large update on Readers and ODE class
I have begun re-implementing the reader with correct doxygen, corrected coding convention. It is only tasked with reading right now: does not treat or check numerical inputs. All of this is saved into a raw data structure which is then used to create an ODE. The ode class takes care of all parsing, converting and whatever is needed.
Missing for now: implementation of how to parse for functions and initial condition y0 correctly.
Missing for reader: I only created the reader class for csv now. there is a strict format of giving the first column as variable names, and the second column with the values that we want  for the different items to take. For the time being, there is no check on parameters that must be provided in the reader, as I said, it purely reads. The ODE will need to verify that essential information is provided or not.

Implementing the other readers is not going to be difficult I think, but how everything is parsed and saved in ODE still requires work. If you can, please go over the headers at least, to see if its ok what I am doing. We can discuss about it on wednesday, or sooner if you want to/are available at one point.  I will need to know how the funciton class is implemented so that I can address it properly in the reader and in the Ode class. I am not sure about the methods to do this properly...