wednesday discussion
reader to be done by friday.
reader does all parsing.
ODE essentially serves as a storage for the problem. Does not contain the numerical parameters that are needed for the solver.

ODE does not contain the solution either.

Reader will be limited to reading from files for now.

ODE can be contructed by code as well, no need to go through Reader constructor.